+++ src/lib.rs
@@ -25,13 +25,17 @@
 //! ## Usage
 //!
 //! This crate provides high-level, idiomatic Rust bindings for GDAL.
-//! To do that, it uses [`gdal_sys`] internally, a low-level interface to the GDAL C library,
-//! which is generated using [`bindgen`](https://rust-lang.github.io/rust-bindgen/).
-//! Using the `gdal-sys` crate directly is normally not needed, but it can be useful in order to call APIs that have not yet been exposed in this crate.
+//! To do that, it uses [`gdal-sys`](gdal-sys) internally, a low-level interface to the GDAL C library, which is generated using [`bindgen`](https://rust-lang.github.io/rust-bindgen/).
+//! Using the `gdal-sys` crate directly is normally not needed, but it can be useful in order to call APIs that have not yet been exposed in `gdal`.
+//!
+//! ## Version support
+//!
+//! As a general rule, only GDAL versions in Ubuntu LTS-1 (previous LTS version, that is, GDAL 3.0 in 20.04 at this moment) are supported.
+//! `gdal-sys` might support earlier versions using the `bindgen` feature flag, but `gdal` does not.
 //!
 //! Building this crate assumes a compatible version of GDAL is installed with the corresponding header files and shared libraries.
-//! This repository includes pre-generated bindings for GDAL 2.4 through 3.5 (see the`gdal-sys/prebuilt-bindings` directory).
-//! If you're compiling against a later version of GDAL, you can enable the `bindgen` feature flag to have new bindings generated on the fly.
+//! This repository includes pre-generated bindings for GDAL 3.0 through 3.8 (see the`gdal-sys/prebuilt-bindings` directory).
+//! If you're compiling against another version of GDAL, you can enable the `bindgen` feature flag to have the bindings generated on the fly.
 //!
 //! ## Show Me Code!
 //!